groq/requirements.txt
groq
python-dotenv
instructor
pydantic